+++ packages/cohere/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/cohere
 
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies [886e7cd]
+  - @ai-sdk/provider-utils@3.0.7
+
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies [1b5a3d3]
+  - @ai-sdk/provider-utils@3.0.6
+
 ## 2.0.5
 
 ### Patch Changes